refactor: move RelayResult and BroadcastStatus to commons
Move RelayResult sealed class and BroadcastStatus enum from
amethyst.service.broadcast to commons.service.broadcast for KMP.
Rename BroadcastModels.kt to BroadcastEvent.kt (now single class).
Update all consumers to import from commons directly.
