+++ amethyst/src/main/java/com/vitorpamplona/amethyst/service/broadcast/BroadcastEvent.kt
@@ -21,48 +21,10 @@
 package com.vitorpamplona.amethyst.service.broadcast
 
 import androidx.compose.runtime.Immutable
+import com.vitorpamplona.amethyst.commons.service.broadcast.BroadcastStatus
+import com.vitorpamplona.amethyst.commons.service.broadcast.RelayResult
 import com.vitorpamplona.quartz.nip01Core.core.Event
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-
-/**
- * Result of a relay's response to an event publish.
- */
-@Immutable
-sealed class RelayResult {
-    /** Relay accepted the event (OK message with success=true) */
-    data object Success : RelayResult()
-
-    /** Relay rejected the event (OK message with success=false) */
-    data class Error(
-        val message: String,
-    ) : RelayResult()
-
-    /** Relay did not respond within timeout */
-    data object Timeout : RelayResult()
-
-    /** Waiting for relay response */
-    data object Pending : RelayResult()
-
-    /** Retry in progress for this relay */
-    data object Retrying : RelayResult()
-}
-
-/**
- * Overall status of a broadcast operation.
- */
-enum class BroadcastStatus {
-    /** Currently waiting for relay responses */
-    IN_PROGRESS,
-
-    /** All relays accepted the event */
-    SUCCESS,
-
-    /** Some relays accepted, some failed */
-    PARTIAL,
-
-    /** No relays accepted the event */
-    FAILED,
-}
 
 /**
  * Tracks a single event broadcast to multiple relays.
